fix: resolve the sticky footer issue
This commit resolves the "sticky footer" problem. On pages with short content (like this sign-up form), the footer renders right after the content, leaving a gap at the bottom of the screen.
"flex-grow: 1;" takes all the empty space(stretches to fill) in the body section, thus pushing the footer to the bottom.